scripts/requirements.txt
# HumanAid Data Collection Scripts - Python Dependencies

# Google Places API
googlemaps==4.10.0

# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2
lxml==4.9.3

# Data Processing
pandas==2.1.3
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9

# Utilities
tqdm==4.66.1
